# openapi.yaml
openapi: 3.0.0
info:
  title: 'MedArea RESTful API'
  version: 1.0.5
paths:
...
  /api/files:
    get:
      tags:
        - file
        - api
      description: 'Get files by filters.'
      operationId: apiFilesIndex
      parameters:
        -
          name: 'user_ids[]'
          in: query
          description: 'Array of user IDs'
          required: true
          schema:
            type: array
            items:
              type: string
        -
          name: 'ids[]'
          in: query
          description: 'Array of files IDs'
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: 'Files data.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResourceCollection'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '422':
          description: 'Validation error.'
    post:
      tags:
        - file
        - api
      description: 'Add files for user.'
      operationId: apiFilesUpload
      requestBody:
        $ref: '#/components/requestBodies/UploadFilesRequest'
      responses:
        '201':
          description: 'Files for user added.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResourceCollection'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '422':
          description: 'Validation error.'
    delete:
      tags:
        - file
        - api
      description: 'Delete files by filters.'
      operationId: apiFilesDestroy
      parameters:
        -
          name: 'user_ids[]'
          in: query
          description: 'Array of user IDs'
          required: true
          schema:
            type: array
            items:
              type: string
        -
          name: 'ids[]'
          in: query
          description: 'Array of files IDs'
          schema:
            type: array
            items:
              type: string
      responses:
        '204':
          description: 'Files deleted.'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '422':
          description: 'Validation error.'
...
components:
  schemas:
...
    FileResource:
      title: FileResource
      properties:
        id:
          description: 'id of File'
          type: string
          format: uuid
        user_id:
          description: 'user_id of File'
          type: string
          format: uuid
        path:
          description: 'path of File'
          type: string
      type: object
    FileResourceCollection:
      title: FileResourceCollection
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/FileResource'
      type: object
...
  requestBodies:
...
    UploadFilesRequest:
      description: 'Upload files request.'
      content:
        multipart/form-data:
          schema:
            required:
              - user_id
              - files
            properties:
              user_id:
                description: 'User ID'
                type: integer
                format: int64
              files:
                description: 'List of files.'
                type: array
                items:
                  type: string
                  format: binary
            type: object
...